Revert "마이바티스 오류나서 고객센터 작업중 커밋 올립니다."
