--- src/main/java/com/kh/klib/customer/exception/CustomerException.java
@@ -1,13 +1,0 @@
-package com.kh.klib.customer.exception;
-
-public class CustomerException extends Exception {
-	
-	public CustomerException() {
-		
-	}
-	
-	public CustomerException(String msg) {
-		super(msg);
-	}
-
-}
